pyproject.toml
[project]
name = "biometric-attendance-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bidict==0.22.1",
    "blinker==1.7.0",
    "click==8.1.7",
    "Flask==3.0.0",
    "Flask-Login==0.6.3",
    "Flask-SocketIO==5.3.6",
    "Flask-SQLAlchemy==3.1.1",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "itsdangerous==2.1.2",
    "Jinja2==3.1.2",
    "MarkupSafe==2.1.3",
    "numpy==1.26.3",
    "pandas==2.1.4",
    "python-dateutil==2.8.2",
    "python-engineio==4.8.1",
    "python-socketio==5.10.0",
    "pytz==2023.3.post1",
    "schedule==1.2.1",
    "simple-websocket==1.0.0",
    "six==1.16.0",
    "SQLAlchemy==2.0.25",
    "typing_extensions==4.9.0",
    "tzdata==2023.4",
    "Werkzeug==3.0.1",
    "wsproto==1.2.0",
    "openpyxl",
    "twilio",
    "celery==5.3.6",
    "apscheduler",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "cryptography>=45.0.4",
]
